backend/requirements.txt
django==3.2
djangorestframework==3.12.4
flake8==6.0.0
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
Pillow==9.3.0
django-filter==22.1